requirements.txt
fastapi
uvicorn
sqlmodel
google-generativeai
pydantic
python-dotenv
psycopg2-binary

passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
